notifications fixed, um erro estranho ta rolando ainda
